backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
pandas==2.1.1
openpyxl==3.1.2
Pillow==10.0.1
pytesseract==0.3.10
googlemaps==4.10.0
psycopg2-binary==2.9.9
python-jose==3.3.0
bcrypt==4.0.1 